pyproject.toml
[project]
name = "img-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "attrs>=25.4.0",
    "cachetools>=6.2.2",
    "certifi>=2025.11.12",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "cryptography>=46.0.3",
    "google-auth>=2.43.0",
    "google-genai>=1.52.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "idna>=3.11",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "markdown-it-py>=4.0.0",
    "mcp[cli]>=1.22.0",
    "mdurl>=0.1.2",
    "pillow>=12.0.0",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.23",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "referencing>=0.37.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "rpds-py>=0.29.0",
    "rsa>=4.9.1",
    "shellingham>=1.5.4",
    "sniffio>=1.3.1",
    "sse-starlette>=3.0.3",
    "starlette>=0.50.0",
    "tenacity>=9.1.2",
    "typer>=0.20.0",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "urllib3>=2.5.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
